requirements.txt
pandas
streamlit
pandasai
langchain_groq
python-dotenv